requirements.txt
python-dotenv==0.17.1
python-telegram-bot==13.7
requests==2.23.0
urllib3==1.25.9